Cleancode6 - BFS 알고리즘의 비재귀화 지금까지의 협업 과정에서 13가지의 버그를 발견하여 수정하였습니다. BFS 알고리즘을 최지웅님이 작업해두셨던 changeBFS 브랜치를 이용하여 비재귀 알고리즘으로 수정하였습니다.
이제 기능적으로 게임이 완벽하게(원본과 똑같이) 구동이 가능합니다.
