+++ src/main.c
@@ -3,16 +3,20 @@
 #include <string.h>
 #include <time.h>
 
-int len, col, num, seed, visi = 0, init = 0;
-// Length，Column，Mines, Seed, Visible Grid, Initialized or not
-unsigned char *m;
-// 00000000
-//        ^ Is mine or not
-//       ^  Is visible or not
-//     ^~   Mark and type
-// ^~~~     Mines number around this grid
-char p[] = { ' ', 'O', 'X', '_' };
-// Mark characters
+/* 구조체 정의 */
+typedef struct AreaInfo { //지뢰판의 구역의 정보를 다루는 구조체
+	int isMine;       //지뢰의 유무를 저장하는 변수
+	int isVisible;    //해당 칸이 보여지는 상태를 저장하는 변수
+	int mark;         //해당 칸의 표식이 무엇인지 저장하는 변수
+	int mineNum;      //해당 칸의 주변 지뢰의 수를 저장하는 변수
+}AreaInfo;
+
+typedef struct SetInfo { // 게임 시작시에 설정하는 맵의 길이, 높이, 지뢰의 갯수, 시드를 저장하는 구조체
+	int len;          // 맵의 길이를 저장하는 변수
+	int col;		  // 맵의 높이를 저장하는 변수
+	int num;          // 지뢰의 갯수를 저장하는 변수
+	int seed;         // 시드를 저장하는 변수
+}SetInfo;
 
 typedef struct Pos { //계산 혹은 함수값 리턴시 x, y의 값을 한번에 다루기 위한 구조체
 	int posX;
@@ -24,25 +28,138 @@ typedef struct StackNode { //스택 구조를 사용하기 위한 구조체
 	struct StackNode *next;
 }StackNode;
 
-#define NONE 0
-#define SAFE 1
-#define MINE 2
-#define QUERY 3
+/* 전역 변수 선언 */
+//부울 연산과 정수 연산을 시각적으로 구분하기 위한 상수
+//코드 여러 군데에서 빈번하게 사용되기 때문에 전역 변수로 선언하였습니다
+//C 문법만 사용 옵션(C++ 문법 제외)을 넣은 Visual Studio 상에서 true, false, bool, boolean 키워드가 작동하지 않아 선언했습니다.
+const int TRUE = 1;	//참인 상태를 나타낼때 쓰일 상수로, 정수 1의 값을 가진다.
+const int FALSE = 0;	//거짓인 상태를 나타낼때 쓰일 상수로, 정수 0의 값을 가진다.
+AreaInfo **areaInfo = NULL; //지뢰판의 정보를 저장할 변수
 
-#define M(x, y) m[(x)*col+(y)]
-#define IS_MINE(x, y) (M(x, y)&1)
-#define IS_VISI(x, y) (M(x, y)&2)
-// Visible or not
-#define IS_MARK(x, y) (M(x, y)&12)
-#define IS_NUM(x, y) (M(x, y)&240)
-#define IS_OUT(x, y) (x>=len||y>=col||x<0||y<0)
-// Out of range or notF
-#define GET_MARK(x, y) ((M(x, y)&12)>>2)
-#define GET_NUM(x, y) ((M(x, y)&240)>>4)
-#define SET_MINE(x, y, s) M(x, y)=s?M(x, y)|1:M(x, y)&254
-#define SET_VISI(x, y, s) M(x, y)=s?M(x, y)|2:M(x, y)&254
-#define SET_MARK(x, y, s) M(x, y)=s&2?s&1?M(x, y)|12:(M(x, y)&243)|8:s&1?M(x, y)&243|4:M(x, y)&243
-#define INC_NUM(x, y) if (!IS_OUT(x, y)) M(x, y)=(M(x, y)&15)|(GET_NUM(x, y)+1)<<4
+/* 함수 원형 선언 */
+//todo: 맨 마지막에 작업할 예정입니다. (디버깅시 혼동 방지를 위해)
+
+//해당 좌표의 지뢰 유무 값 전달
+int IsHereMine(int x, int y) {
+	int result;
+	result = areaInfo[x][y].isMine;
+	return result;
+}
+
+//해당 좌표가 보여지는 상태에 대한 값 전달
+int IsVisible(int x, int y) {
+	int result;
+	result = areaInfo[x][y].isVisible;
+	return result;
+}
+
+//해당 좌표의 표식의 값 전달
+int Mark(int x, int y) {
+	int result;
+	result = areaInfo[x][y].mark;
+	return result;
+}
+
+//해당 칸 주변 지뢰의 수 전달
+int MineNum(int x, int y) {
+	int result;
+	result = areaInfo[x][y].mineNum;
+	return result;
+}
+
+//입력된 좌표가 범위 밖인지 판단
+int IsOut(int x, int y, SetInfo setInfo) {
+	const int OUT = 1; //입력 좌표가 범위 밖임을 나타내는 상수, 정수 1의 값을 가진다.
+	const int IN = 0; //입력 좌표가 범위 안에 있음을 나타내는 상수, 정수 0의 값을 가진다.
+
+	if (x >= setInfo.len || y >= setInfo.col || x < 0 || y < 0) {
+		return OUT;
+	}
+	return IN;
+}
+
+void SetMine(int x, int y) {
+	areaInfo[x][y].isMine = 1;
+	return;
+}
+
+//해당 칸을 볼 수 있는지 아닌지에 대한 값 설정
+void SetVisible(int x, int y, int s) {
+	if (s == TRUE) {
+		areaInfo[x][y].isVisible = 1;
+	}
+	if (s == FALSE) {
+		areaInfo[x][y].isVisible = 0;
+	}
+	return;
+}
+
+//마크 값 설정
+void SetMark(int x, int y, int s) {
+	switch (s) {
+	case 0:
+		areaInfo[x][y].mark = 0;
+		break;
+
+	case 1:
+		areaInfo[x][y].mark = 1;
+		break;
+
+	case 2:
+		areaInfo[x][y].mark = 2;
+		break;
+
+	case 3:
+		areaInfo[x][y].mark = 3;
+		break;
+
+	default:
+		return;
+	}
+}
+
+void IncNum(int x, int y, SetInfo setInfo) {
+	if (IsOut(x, y, setInfo))
+		return;
+	else
+		areaInfo[x][y].mineNum = areaInfo[x][y].mineNum + 1;
+	return;
+}
+
+void InitArea(SetInfo setInfo) {
+	/*
+	기능: 지뢰판을 이차원 배열로 할당하고 초기화가 필요한 구조체 멤버들을 초기화한다.
+	파라미터 : len은 지뢰판의 행의 수를 col은 지뢰판의 열의 수를 나타낸다.
+	*/
+	const int ZERO = 0;  //갯수가 0일때 쓰일 상수로, 정수 0의 값을 가진다.
+	int i;
+	int j;
+
+	//변수 areaInfo에 이차원 배열을 할당한다.
+	areaInfo = (AreaInfo **)malloc(setInfo.len * sizeof(AreaInfo *));
+	if (areaInfo == NULL) { // 동적 할당에 실패한 경우 아래 문구와 함께 종료된다.
+		printf("Not enough memory!");
+		return;
+	}
+	for (i = 0; i < setInfo.len; i++) {
+		areaInfo[i] = (AreaInfo *)malloc(setInfo.col * sizeof(AreaInfo)); // 동적 할당에 실패한 경우 아래 문구와 함께 종료된다.
+		if (areaInfo[i] == NULL) {
+			printf("Not enough memory!");
+			return;
+		}
+	}
+	
+	//구조체 배열의 멤버 값을 초기화한다.
+	for (i = 0; i < setInfo.len; i++) {
+		for (j = 0; j < setInfo.col; j++) {
+			areaInfo[i][j].isMine = FALSE;
+			areaInfo[i][j].isVisible = FALSE;
+			areaInfo[i][j].mark = ZERO;
+			areaInfo[i][j].mineNum = ZERO;
+		}
+	}
+	return;
+}
 
 StackNode *Push(StackNode *stack, Pos pos) { //스택에 푸시하는 함수
 	StackNode *newNode; //새로 생성한 노드
@@ -55,7 +172,7 @@ StackNode *Push(StackNode *stack, Pos pos) { //스택에 푸시하는 함수
 
 	if (stack == NULL) //스택이 비어있다면 새로 생성한 노드를 리턴
 		return newNode;
-	
+
 	pushNode = stack; //스택이 비어있지않다면 현재의 마지막 노드로 이동하여 새로 생성한 노드를 마지막에 오도록 추가
 	while (pushNode->next != NULL)
 		pushNode = pushNode->next;
@@ -69,18 +186,15 @@ Pos Pop(StackNode **stack) { //스택의 노드를 팝하는 함수
 	StackNode *popNode; //메모리 해제를 하기 위해 팝할 노드를 가리킬 변수
 
 	e = (*stack)->ele; //팝할 노드의 원소와 노드를 저장
-	popNode = *stack; 
+	popNode = *stack;
 
 	*stack = (*stack)->next; //스택이 스택의 다음 노드를 가리키게 함
-	
+
 	free(popNode); //팝한 노드 메모리 할당 해제
 	return e; //팝한 노드의 원소를 리턴
 }
 
 int IsEmptyStack(StackNode *stack) { //스택이 비어있는지 확인하는 함수
-	const int TRUE = 1; 
-	const int FALSE = 0;
-
 	if (stack == NULL) //스택이 비어있다면 TRUE를 리턴
 		return TRUE;
 	return FALSE; //스택이 비어있지않다면 FALSE를 리턴
@@ -110,17 +224,17 @@ void SetAdjecantPosition(Pos *adjPos, Pos pos) { //x, y 주변의 인접한 8칸
 
 	adjPos[7].posX = pos.posX - 1; //북서쪽에 있는 좌표
 	adjPos[7].posY = pos.posY + 1;
-	
+
 	return;
 }
 
-void bfs(int x, int y) {
+void bfs(SetInfo setInfo, int x, int y, int* visi) {
 	const int adjNum = 8; //인접한 칸의 최대 갯수
 	StackNode *stack = NULL; //스택을 가리킬 헤드 포인터
 	Pos adjPos[8]; //스택에서 팝한 좌표의 주변 8칸의 좌표값을 저장할 배열
 	Pos pos; //팝한 노드의 좌표를 저장할 변수
 
-	if (IS_OUT(x, y) || IS_VISI(x, y)) //선택한 좌표가 보이는 상태이거나 범위 밖의 좌표이면 리턴
+	if (IsOut(x, y, setInfo) || IsVisible(x, y)) //선택한 좌표가 보이는 상태이거나 범위 밖의 좌표이면 리턴
 		return;
 
 	pos.posX = x; //선택한 노드의 x, y 좌표를 구조체 변수에 저장
@@ -130,15 +244,14 @@ void bfs(int x, int y) {
 
 	while (!IsEmptyStack(stack)) { //스택이 비어있지않다면 시행
 		pos = Pop(&stack); //스택의 노드를 팝하고 그 원소를 저장
-		
-		if (!(IS_OUT(pos.posX, pos.posY) || IS_VISI(pos.posX, pos.posY))) { //팝한 좌표가 보이는 상태와 범위 밖의 좌표가 아닐 경우 실행
-			++visi; //보이는 수를 증가시키고 해당 좌표의 칸을 보이는 상태로 변경
-			SET_VISI(pos.posX, pos.posY, 1);
-			
-			if (!GET_NUM(pos.posX, pos.posY)) { //주변의 지뢰가 매설되어 있지 않은 경우 실행
-				SetAdjecantPosition(adjPos, pos); //주변 8칸의 좌표를 배열에 저장
 
-				for (int i = 0;i < adjNum;i++) { //8칸의 좌표를 스택에 푸시
+		if (!(IsOut(pos.posX, pos.posY, setInfo) || IsVisible(pos.posX, pos.posY))) { //팝한 좌표가 보이는 상태와 범위 밖의 좌표가 아닐 경우 실행
+			++(*visi); //보이는 수를 증가시키고 해당 좌표의 칸을 보이는 상태로 변경
+			SetVisible(pos.posX, pos.posY, 1);
+
+			if (!(areaInfo[pos.posX][pos.posY].mineNum)) { //주변의 지뢰가 매설되어 있지 않은 경우 실행
+				SetAdjecantPosition(adjPos, pos); //주변 8칸의 좌표를 배열에 저장
+				for (int i = 0; i < adjNum; i++) { //8칸의 좌표를 스택에 푸시
 					stack = Push(stack, adjPos[i]);
 				}
 			}
@@ -148,38 +261,46 @@ void bfs(int x, int y) {
 	return;
 }
 
-void print(int c) {
+void print(int c, SetInfo setInfo) {
+	char p[] = { ' ', 'O', 'X', '_' };
 	int x, y;
+	char mineNumToChar;
+
 	printf("  ");
-	for (x = 0; x < len; ++x) {
+	for (x = 0; x < setInfo.len; ++x) {
 		printf("%-2d", x);
 	}
 	printf("\n");
-	for (y = col - 1; y >= 0; --y) {
+	for (y = setInfo.col - 1; y >= 0; --y) {
 		printf("%-2d", y);
 		if (c)
-			for (x = 0; x < len; ++x) {
-				if (IS_MINE(x, y))
+			for (x = 0; x < setInfo.len; ++x) {
+				if (IsHereMine(x, y))
 					printf("%-2c", '*');
-				else
-					printf("%-2c", GET_NUM(x, y) + '0');
+				else {
+					mineNumToChar = (char)areaInfo[x][y].mineNum + '0';
+					printf("%-2c", mineNumToChar);
+				}
 			}
 		else
-			for (x = 0; x < len; ++x) {
-				if (IS_VISI(x, y)) {
-					if (IS_MINE(x, y))
+			for (x = 0; x < setInfo.len; ++x) {
+				if (IsVisible(x, y)) {
+					if (IsHereMine(x, y))
 						printf("%-2c", '*');
-					else
-						printf("%-2c", GET_NUM(x, y) + '0');
+					else {
+						mineNumToChar = (char)areaInfo[x][y].mineNum + '0';
+						printf("%-2c", mineNumToChar);
+					}
+						
 				}
 				else
-					printf("%-2c", p[GET_MARK(x, y)]);
+					printf("%-2c", p[Mark(x, y)]);
 			}
 		printf("%-2d", y);
 		printf("\n");
 	}
 	printf("  ");
-	for (x = 0; x < len; ++x) {
+	for (x = 0; x < setInfo.len; ++x) {
 		printf("%-2d", x);
 	}
 	printf("\n");
@@ -187,117 +308,137 @@ void print(int c) {
 		printf("WARNING: CHAETING DETECTED\n");
 }
 
-void clear(void) {
-	for (int i = 0; i < len*col; ++i)
-		m[i] |= 2;
+void Gameover(SetInfo setInfo) {
+	int i, j;
+	for (i = 0; i < setInfo.len; ++i) {
+		for (j = 0; j < (setInfo.col); j++) {
+			areaInfo[i][j].isVisible = 1;
+		}
+	}
+	return;
 }
 
-int input(void) {	//사용자의 입력, 입력값 검증, 지뢰 탐색 등의 여러 기능이 있는 함수.
+int input(SetInfo setInfo, int* init, int* visi) {	//사용자의 입력, 입력값 검증, 지뢰 탐색 등의 여러 기능이 있는 함수.
 	//TODO: 함수 기능 세분화(하나의 함수가 하나의 기능만 하도록 하기) 하기.
 	int x, y, s;	//사용자로부터 입력받을 x,y좌표와 명령어(s)
+	int a, b, i;	//랜덤으로 선정된 x,y좌표쌍 (a,b)와 for문 반복자
+	int leftMineToInstall = setInfo.num;
 
 	//사용자 입력 처리
 	printf("Enter X coordinate, Y coordinate, and instruction\n");
 	scanf("%d %d %d", &x, &y, &s);
 
 	//입력값 검증 (좌표값 범위 검사)
-	if (IS_OUT(x, y)) {
-		print(0);
+	if (IsOut(x, y, setInfo)) {
+		print(0, setInfo);
 		printf("Invaild Command: Out of range\n");
 		return 0;	//게임 계속 진행
 	}
 
 	//게임판 초기화
-	if (!init) {
-		for (int i = 0, a, b; i < num; ++i) {
+	if (*init != TRUE) {
+		for (i = 0; i < leftMineToInstall; ++i) {
 			//반복해서 랜덤한 (a, b) 좌표 생성
-			a = rand() % len;
-			b = rand() % col;
-			if (IS_MINE(a, b) || (a == x && b == y))
+			a = rand() % setInfo.len;
+			b = rand() % setInfo.col;
+			if (IsHereMine(a, b) || (a == x && b == y)) {
+				leftMineToInstall++;	//이번 반복횟수를 무효로 하기 위해 1 증가시켜야 함. (continue 하더라도 i는 1 증가해버림. 이를 되돌려주기 위한 조치)
 				continue;
+			}
 			else {
 				//(a,b)좌표에 지뢰 매설 및 주변 8개 칸의 안내숫자 1씩 증가
-				INC_NUM(a - 1, b - 1);
-				INC_NUM(a - 1, b);
-				INC_NUM(a - 1, b + 1);
-				INC_NUM(a, b - 1);
-				SET_MINE(a, b, 1);
-				INC_NUM(a, b + 1);
-				INC_NUM(a + 1, b - 1);
-				INC_NUM(a + 1, b);
-				INC_NUM(a + 1, b + 1);
+				IncNum(a - 1, b - 1, setInfo);
+				IncNum(a - 1, b, setInfo);
+				IncNum(a - 1, b + 1, setInfo);
+				IncNum(a, b - 1, setInfo);
+				SetMine(a, b);
+				IncNum(a, b + 1, setInfo);
+				IncNum(a + 1, b - 1, setInfo);
+				IncNum(a + 1, b, setInfo);
+				IncNum(a + 1, b + 1, setInfo);
 			}
 		}
-		init = 1;	//초기화 완료
+		*init = TRUE;	//초기화 완료
 	}
 
 	//명령어 검사 및 분기처리
 	if (s) {
-		if (s == -seed) {	//-seed 값을 명령어로 입력할 경우 치트 동작
-			print(1);
+		if (s == (-1)*(setInfo.seed)) {	//-seed 값을 명령어로 입력할 경우 치트 동작
+			print(1, setInfo);
 		}
 		else if (s <= 4 && s > 0) {
 			//해당 좌표에 메모하는 명령(1,2,3,4)를 입력했을 경우
-			if (!IS_VISI(x, y))
-				SET_MARK(x, y, s - 1);
+			if (!IsVisible(x, y))
+				SetMark(x, y, s - 1);
 			else {
-				print(0);
+				print(0, setInfo);
 				printf("Invaild Command: Already visible\n");
 				return 0;	//게임 계속 진행
 			}
-			print(0);
+			print(0, setInfo);
 		}
 		else {
 			//잘못된 명령어를 입력했을 경우
-			print(0);
+			print(0, setInfo);
 			printf("Invaild Command: Command does not exist\n");
 		}
 	}
 	else {
 		//지뢰를 파보는 명령(0)을 입력했을 경우
-		SET_MARK(x, y, 0);
-		if (IS_MINE(x, y)) {	//지뢰 밟았을 경우 패배 처리
+		SetMark(x, y, 0);
+		if (IsHereMine(x, y)) {	//지뢰 밟았을 경우 패배 처리
 			printf("You have lost\n");
-			clear();
-			print(0);
+			Gameover(setInfo);
+			print(0, setInfo);
 			return 1;	//패배
 		}
-		bfs(x, y);
-		print(0);
+		bfs(setInfo, x, y, visi);
+		print(0, setInfo);
 	}
-	if (visi == len * col - num) {	//남은 지뢰 갯수 검사
+	if (*visi == (setInfo.len) * (setInfo.col) - (setInfo.num)) {	//남은 지뢰 갯수 검사
 		//다 찾았을 경우 승리 처리
 		printf("You win\n");
-		clear();
-		print(0);
+		Gameover(setInfo);
+		print(0, setInfo);
 		return 2;	//승리
 	}
 	return 0;	//게임 계속 진행
 }
 
 int main(int argc, char **argv) {
+	SetInfo setInfo;
+	int visi = 0;	//밝혀진 지뢰의 숫자를 0으로 초기화한다
+	int init = FALSE;	//아직 지뢰판 초기화가 되지 않았으므로 초기화 여부를 FALSE로 설정한다
+	int i = 0;
+
 	if (argc > 3) {
 		printf("Getting information from argument\n");
-		len = atoi(argv[1]);
-		col = atoi(argv[2]);
-		num = atoi(argv[3]);
+		setInfo.len = atoi(argv[1]);
+		setInfo.col = atoi(argv[2]);
+		setInfo.num = atoi(argv[3]);
 		if (argc > 4)
-			seed = atoi(argv[4]);
-		else seed = (int)time(NULL);
+			setInfo.seed = atoi(argv[4]);
+		else setInfo.seed = (int)time(NULL);
 	}
 	else {
 		printf("Enter length, width, mineage, and seed(random seeds fill in -1)\n");
-		scanf("%d %d %d %d", &len, &col, &num, &seed);
-		if (seed < 0)
-			seed = (int)time(NULL);
+		scanf("%d %d %d %d", &setInfo.len, &setInfo.col, &setInfo.num, &setInfo.seed);
+		if (setInfo.seed < 0)
+			setInfo.seed = (int)time(NULL);
 	}
-	srand(seed);
-	printf("Length: %d\n", len);
-	printf("Column: %d\n", col);
-	printf("Mines: %d\n", num);
-	printf("Seed：%d\n", seed);
-	m = (unsigned char *)malloc(len*col);
-	memset(m, 0, len*col);
-	print(0);
-	for (; !input(););
+
+	srand(setInfo.seed);
+	printf("Length: %d\n", setInfo.len);
+	printf("Column: %d\n", setInfo.col);
+	printf("Mines: %d\n", setInfo.num);
+	printf("Seed：%d\n", setInfo.seed);
+	InitArea(setInfo);
+	print(0, setInfo);
+	for (; !input(setInfo, &init, &visi););
+
+	for (i = 0; i < setInfo.len; i++) {
+		free(areaInfo[i]);
+	}
+
+	free(areaInfo);
 }
